app/requirements.txt
black==23.7.0
isort==5.12.0
# pytest==7.4.0
ruff==0.0.278
scikit-image==0.21.0
numpy==1.24.4
ffmpeg-python==0.2.0
opencv-python-headless==4.7.0.72
python-dotenv==1.0.0
uvicorn[standard]==0.22.0
fastapi==0.100.0